Move templates to pkg/templates
